feat(schedule): refactor search attributes inputs
Refactored the search attributes inputs into a new reusable component
(SchedulesSearchAttributesInputs). Removed redundant imports and
replaced the Tabs implementation in schedule-form-view.svelte with the
new component.
